DEV: created metrics module with refactored routines that produced Pandas databases
