+++ pyfibre/io/object_io.py
@@ -1,10 +1,8 @@
-from pyfibre.io.utils import get_networkx_graph
 from pyfibre.model.objects.base_graph_segment import BaseGraphSegment
 from pyfibre.model.objects.cell import Cell
 from pyfibre.model.objects.fibre import Fibre
 from pyfibre.model.objects.fibre_network import FibreNetwork
-from pyfibre.utilities import (
-    save_pickle, load_pickle, save_json, load_json)
+from pyfibre.utilities import save_json, load_json
 
 from .segment_io import save_segments, load_segments
 
@@ -100,32 +98,3 @@ def load_fibre_networks(file_name, image=None):
     """Load a list of FibreNetwork instances"""
     return load_base_graph_segments(
         file_name, 'fibre_networks', FibreNetwork, image=image)
-
-
-def save_objects(objects, file_name, file_type=None):
-    """Loads pickled image objects"""
-
-    if file_type is not None:
-        file_name = '_'.join([file_name, file_type])
-
-    try:
-        save_pickle(objects, f"{file_name}.pkl")
-    except IOError as e:
-        raise IOError(
-            f"Cannot save to file {file_name}.pkl"
-        ) from e
-
-
-def load_objects(file_name, file_type=None):
-    """Loads pickled image objects"""
-
-    if file_type is not None:
-        file_name = '_'.join([file_name, file_type])
-
-    try:
-        objects = load_pickle(f"{file_name}.pkl")
-        return objects
-    except IOError as e:
-        raise IOError(
-            f"Cannot read file {file_name}.pkl"
-        ) from e
