+++ pyfibre/model/objects/fibre.py
@@ -1,8 +1,7 @@
 import pandas as pd
-import networkx as nx
 import numpy as np
 
-from pyfibre.model.tools.analysis import segment_analysis
+from pyfibre.model.tools.metrics import segment_metrics
 from pyfibre.model.tools.fibre_utilities import branch_angles
 
 from .base_graph_segment import BaseGraphSegment
@@ -80,13 +79,13 @@ class Fibre(BaseGraphSegment):
         database['Fibre Angle'] = self.angle
 
         if image is not None:
-            segment_metrics = segment_analysis(
+            metrics = segment_metrics(
                 self.segment, image=image, tag='Fibre')
 
         else:
-            segment_metrics = segment_analysis(
+            metrics = segment_metrics(
                 self.segment, tag='Fibre')
 
-        database = database.append(segment_metrics, ignore_index=False)
+        database = database.append(metrics, ignore_index=False)
 
         return database
