+++ pyfibre/model/tools/tests/test_analysis.py
@@ -2,11 +2,8 @@ from unittest import TestCase
 import numpy as np
 
 from pyfibre.model.tools.analysis import (
-    tensor_analysis, angle_analysis, fibre_analysis,
-    fibre_network_analysis
+    tensor_analysis, angle_analysis
 )
-from pyfibre.tests.probe_classes import (
-    ProbeFibre, ProbeFibreNetwork)
 
 
 class TestAnalysis(TestCase):
@@ -36,23 +33,3 @@ class TestAnalysis(TestCase):
         angle_sdi, _ = angle_analysis(angles, weights, n_bin=10)
 
         self.assertEqual(angle_sdi, 0.225)
-
-    def test_fibre_analysis(self):
-
-        tot_fibres = [ProbeFibre(), ProbeFibre(), ProbeFibre()]
-
-        fibre_database = fibre_analysis(tot_fibres)
-        self.assertEqual(3, len(fibre_database))
-        self.assertEqual(11, len(fibre_database.columns))
-
-    def test_fibre_network_analysis(self):
-
-        fibre_network = ProbeFibreNetwork()
-        fibre_network.fibres = fibre_network.generate_fibres()
-        image = np.ones((5, 5))
-        image[2:, 2:] = 2
-
-        fibre_network_metrics = fibre_network_analysis(
-            [fibre_network], image)
-
-        self.assertEqual(15, len(fibre_network_metrics.columns))
